Addressed PR issues
