+++ App.js
@@ -1,29 +1,18 @@
 import { StatusBar } from "expo-status-bar";
 import React from "react";
 import { StyleSheet, View } from "react-native";
-import List from "./components/List/List.View";
-
-const data = [
-  {
-    id: "bd7acbea-c1b1-46c2-aed5-3ad53abb28ba",
-    title: "First Item",
-  },
-  {
-    id: "3ac68afc-c605-48d3-a4f8-fbd91aa97f63",
-    title: "Second Item",
-  },
-  {
-    id: "58694a0f-3da1-471f-bd96-145571e29d72",
-    title: "Third Item",
-  },
-];
+import { SafeAreaProvider as SafeAreaView } from "react-native-safe-area-context";
+import { style } from "./App.style";
+import ListPageDataLayer from "./views/ListPage.dataLayer";
 
 export default function App() {
   return (
-    <View style={styles.container}>
-      <List data={data} />
-      <StatusBar style="auto" />
-    </View>
+    <SafeAreaView style={style.container}>
+      <View style={styles.container}>
+        <ListPageDataLayer />
+        <StatusBar style="auto" />
+      </View>
+    </SafeAreaView>
   );
 }
