+++ App.style.js
@@ -1,0 +1,8 @@
+import { StatusBar, StyleSheet } from "react-native";
+
+export const style = StyleSheet.create({
+  container: {
+    flex: 1,
+    marginTop: StatusBar.currentHeight || 0,
+  },
+});
