+++ components/Header/Header.Style.js
@@ -1,0 +1,8 @@
+import { StyleSheet } from "react-native";
+
+export const style = StyleSheet.create({
+  header: {
+    fontSize: 32,
+    backgroundColor: "#fff",
+  },
+});
