+++ components/Item/Item.Style.js
@@ -1,0 +1,12 @@
+import { StyleSheet } from "react-native";
+
+export const style = StyleSheet.create({
+  item: {
+    padding: 20,
+    marginVertical: 8,
+    marginHorizontal: 16,
+  },
+  title: {
+    fontSize: 32,
+  },
+});
