+++ views/ListPage.js
@@ -1,0 +1,6 @@
+import React from "react";
+import List from "../components/List/List";
+
+export default function ListPage({ data }) {
+  return <List data={data} />;
+}
